backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy>=2.0.35
PyJWT==2.10.1
python-dotenv==1.0.0
openai==1.54.0
SQLAlchemy>=2.0.35
